Remove helpers folder, separate internal bot_ai from BotAI class, merge distances.py into bot_ai_internal.py
